requirements.txt
tqdm
requests
beautifulsoup4
lxml
dspy
pydantic
pandas
fastapi
uvicorn
playwright
supabase
openpyxl
python-dotenv